v1.0.2-SNAPSHOT 更新，更新内容见 readme.txt

+++ xingtool-spring-boot-autoconfigure/src/main/java/top/cutexingluo/tools/designtools/JUC/ThreadPoolProperty.java
@@ -10,6 +10,9 @@ import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 
 /**
+ * 线程池基础数据
+ * <p>配置文件-线程池</p>
+ *
  * @author XingTian
  * @version 1.0.0
  * @date 2023/2/3 16:03
@@ -18,11 +21,32 @@ import java.util.concurrent.TimeUnit;
 @ConfigurationProperties(prefix = "xingtools.xt-thread-pool")
 @Import(ThreadPoolConfiguration.class)
 public class ThreadPoolProperty {
+    /**
+     * 核心线程数
+     */
     private int corePoolSize;
+    /**
+     * 最大线程数
+     */
     private int maxPoolSize;
-    private int keepAliveTime;
+    /**
+     * 活跃时间
+     */
+    private long keepAliveTime;
+    /**
+     * 时间单位
+     */
     private TimeUnit unit;
+    /**
+     * 阻塞队列
+     */
     private BlockingQueue<Runnable> workQueue;
+    /**
+     * 线程工厂
+     */
     private ThreadFactory threadFactory;
+    /**
+     * 拒绝策略
+     */
     private RejectedExecutionHandler handler;
 }
